Move content from end_contract_wizard to sponsorship
