+++ sponsorship_compassion/wizard/__init__.py
@@ -10,3 +10,4 @@
 ##############################################################################
 
 from . import generate_gift_wizard
+from . import end_sponsorship_wizard
